Search for Oauth2Token that are authorized and not invalidated in the Rack middleware
I was surprised that when using `OAuth::Controllers::ApplicationControllerMethods.oauthenticate`, every token were valid for authentication, even the invalidated ones!

So I think this patch fix this (I've updated the specs). I've also simplified a condition in `OAuth::Rack::OAuthFilter#oauth2_token` and the Guardfile to remove the deprecation notice.