OAuthFilter:oauth2_token to rejects headers that explicitly state oauth_version="1.0"
OAuth 1 allows whitespace between the header parameters, but this is optional.  OAuthFilter was using this optionsl behavior to discriminate between oauth1 and oauth2 tokens.  While most OAuth clients put whitespace between their parameters, some clients do not.  The filter was mis-identifying these tokens as OAuth2 tokens, breaking the authentication chain.

oauth2_token now checks if oauth_version="1.0".  If this is present, then the token is rejected.  oauth_version is an optional paramter.  When it is missing, oauth2_token then defaults to checking for whitespace between the parameters.